requirements.txt
Django>=4.2.0
gunicorn==21.2.0
django-multiselectfield>=0.1.13